Add files via upload
Welcome to my project on analyzing video game sales trends for the online store Ice. In this real-world case study, I delved into a dataset spanning back to 2016, with the goal of identifying patterns that determine the success of video games. With access to user and expert reviews, genre classifications, platform information, and historical sales data, the aim was to uncover insights that could inform future advertising campaigns and identify promising projects.